hsmd: remove hsmd/client.c
It was only used by handshake.c.
