+++ channeld/Makefile
@@ -72,6 +72,7 @@ CHANNELD_COMMON_OBJS :=				\
 	common/wire_error.o			\
 	common/wireaddr.o			\
 	gossipd/gen_gossip_wire.o		\
+	hsmd/gen_hsm_client_wire.o		\
 	lightningd/gossip_msg.o			\
 	wire/fromwire.o				\
 	wire/towire.o
