+++ connectd/Makefile
@@ -61,7 +61,6 @@ CONNECTD_COMMON_OBJS :=				\
 	common/wireaddr.o			\
 	common/wire_error.o			\
 	gossipd/gen_gossip_wire.o		\
-	hsmd/client.o				\
 	hsmd/gen_hsm_client_wire.o		\
 	lightningd/gossip_msg.o			\
 	wire/gen_onion_wire.o
