+++ openingd/Makefile
@@ -70,6 +70,7 @@ OPENINGD_COMMON_OBJS :=				\
 	common/wire_error.o			\
 	common/wireaddr.o			\
 	gossipd/gen_gossip_wire.o		\
+	hsmd/gen_hsm_client_wire.o		\
 	lightningd/gossip_msg.o
 
 $(LIGHTNINGD_OPENING_OBJS): $(LIGHTNINGD_HEADERS)
